+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog for zest.releaser
 6.16.1 (unreleased)
 -------------------
 
-- Refuse to edit history header when it may be from an already released version.
-  We look for "unreleased" in it.  Give a warning when this happens.
+- Refuse to edit history header when it looks to be from an already released version.
+  We look for a date in it (like 2019-02-20).  Give a warning when this happens.
   Fixes `issue 311 <https://github.com/zestsoftware/zest.releaser/issues/311>`_.
   [maurits]